wake: make stack tracing optional
It is impossible to support stack tracing without impacting the memory
complexity of programs. Wake can evaluate a function's return before
the arguments to it's body are available. This means all your stack
frames are potentially gone by the time an exception causing expression
is processed.

Once we evaluate in heap order, this problem will be compounded.
Therefore, we provide a mechanism for stack tracing in debug, but leave
it off by default.