database: clean on exit
This does not quite close #16, but it comes close.  We now sweep away all
Jobs that have no impact on the filesystem and whose results are not kept.

The prior database refactor also made the database pro-actively delete
Jobs whose outputs were partially overwritten or whose inputs had their
hash change.

I've elected to intentionally leak a bit of space in the interest of
potentially saving compute later.  It could be that the job which produced
an intermediate build file gets rerun and produces the same outputs.  In
that case, we want to keep any dependent Job results around in the database
so as to skip re-execution.