symbol: fix a precedence parsing bug
lambda and if must have precedence 0 since they introduce a block,
otherwise a '::' operator would be parsed between 'if' and 'then'
but not after an 'else'

':' precedence must be higher than ',' or else type ascription in
a list will break [a,b,c::d] parsing. Note that this problem does
not apply to 'if' or lambda, because they are unary only.

While I'm modifying precedence, promote ` to higher than .
This will let us use it as a tag operator in the future.