fuse: solve a metadata race on out JSON file size
Looking in .fuse/ previously caused the OS' filesystem layer to cache the
size of the output JSON.  This was problematic if further accesses happened
that caused the output JSON file size to change.

Now we let the fuse runner indicate when the job is done and the JSON
should be produced. Once.