build: support libtinfo and unusual gcc install locations
ncurses used to be split in two libraries.
Old versions sometimes did not list the dependency of ncurses on tinfo.
To support old linux, we need to explicitly test for libtinfo.
To support modern systems, we should also try to run without it.

On older systems we sometimes need to use a newer C++ compiler.
Add a WAKE_PATH to the build for controling this location.