match: don't generalize the RHS type of match cases
This was not a bug, per-se, because the types were still correct.

  global def update = match _
    1 = 1
    x = x

However, if you look at the type AST, the type of 'x' here would be free,
because the RHS is the identity function.  Really, we want to force the type
to be Integer, to match the first case.

This change was made to improve the html output experience.
There is a cost in AST complexity, but this should get optimized away later.