String: use strlen + foldl for cat
Getting the evaluation order correct in C++ is quite difficult.
We want to force deferrals of all elements in the list as soon
as the tuple chain is realized. Then, we want to wait for all of
values to become available before we produce the result String.

foldl with strlen just does the right thing automatically.