wake: --no-tty now gathers stdout/stderr of jobs together (#257)
In interactive mode, you want to see progress immediately, so we pass
stdout/stderr from jobs through as it arrives.

However, in batch mode, we don't have this requirement.  When inspecting
logs from a build it's quite difficult to disentangle which error messages
correspond to which command.

To fix this, wake --no-tty will now defer echoing the command run and it's
stderr/stdout until the job is complete.  Then all of the information is
printed together.