FUSE: creating devices/files/dirs/symlinks overwrites invisible files
When a Job takes ownership of a path, we need to clear out any old files in
the underlying workspace that are in the way.

Fixes #318. 'mkdir' now erases an old invisible contents. Therefore, when
a following 'rmdir' arrives, it succeeds. If the 'rmdir' was for a visible
directory (not created by this Job) it will still fail with EACCES.