symbol: process escapes in RegExp more carefully (#341)
Previously, `\\` was a parse error, because the trailing \` was interpreted
as an escaped end of regular expression.  By adding a prior rule to
explicitly detect paired \\s, we only interpret \` when there are an odd
number of prior \s.

Also, `a\`b` previously rendered as RegExp `a`b`, which is clearly wrong.
Since re2 interprets \` as ` already anyway, push \` as \` to re2.
This means everything between ``s is now literally what re2 receives.