fuse: fix a race condition between daemon exit and client attachment
If a client tried to connect to the FUSE daemon between the time when the
FUSE daemon started to exit and when it succeeded in exiting, there was a
race condition, whereby the open of '.f.fuse-waked' succeeds, but the daemon
exist anyway, leaving the client unable to make forward progress.

The solution taken is to recognize that once an exit attempt had been made,
it cannot be stopped.  Instead, any future attempts to open '.f.fuse-waked'
must fail.  Careful synchonization is needed, because canceling the SIGALRM
does not guarantee the absence of invocations of handle_exit.