workspace: stop looking for wake.db if you find .wakeroot
To find the workspace root, wake looks for either a wake.db or a .wakeroot.
It searches from the current directory to the root, parent by parent.

This PR makes wake stop searching when it finds either wake.db or .wakeroot.
Previously, if wake found a wake.db, it would prefer it over a child .wakeroot.
This was a hold-over from when it was actually looking for a wit manifest.

Now that wit is gone, we should trust the .wakeroot to be correct.
There is never a reason to add more scope to the workspace than requested.