wake: don't replace bootstrap gz files by default
For systems without re2c and for the pre-wake bootstrap, we need
copies of the generated parser/lexer files.  When the actual
origin of these files change, the generated code is committed to git.
However, gzip on different systems produces different output.
Therefore, to prevent dirty trees and ping-pong'ing the contents,
only recreate the generated files if UPGRADE_GENERATED=1.