database: add a missing index for foreign key constraint
When a 'stats' table entry is deleted, the jobs table must be checked to
confirm it is unreferenced to ensures the references() is not broken.

Unfortuntely, there was no index, which meant this required a jobs scan.
As the number of stats is O(jobs) and the clean-up phase of a rebuild often
refreshes many of the statistics, this could become O(jobs^2) cost.

By adding this index, deleting a stats row is now O(1).