[json] Draft proposal for new way to use json
This change introduces two things

1) The concept of a Traversal
2) The small family of helpers (viewOnlyT, viewAllT, mapOnlyT, jObjectMap,jObjectF, etc...)

Togethor I give examples of how these compose togethor by showing how to translate jq examples
into wake using this. The TLDR is the following

A traversal is a way of describing how to "access" many parts of an arbitrary data structure at once.

You can edit all the members of a traversal or you can get a list of them. They all have to be the
same type but thats about it. This is really useful for accessing highly nested structures and
either reading values out of them or modifying the strcture as if it were an inplace modification.

Traversals ensure the following:

editT s id = s
viewT (editT s f) t = map f (viewT s)

This means that they can't be used for making new data in certain ways. Traversals can only express
paths within data structures and not modifications to them but you can use editT to specify that
all of those locations should be modified however.

Traversals are being used as a bit of a one size fits all kind of thing but I've provided the following
helpers for viewing their results and error handeling:

viewOnlyT - returns its error value if the traversal is not a singelton
viewFirstT - returns its error value if the result is empty or the first element of the traversal otherwise
viewAllT - no possible error, just wraps viewT in a Pass so that it composes well in cases like jObjectMap

In order to have slightly better composition of more complicated things like creating a new object,
I've introduced things like jObjectF and jObjectMap. These have preety complicated types but they
reflect how jq performs with the additional detail that they allow explicit error handeling.