[json] Use applicitive |> and $ to compose Setters/Getters seperately
Have a long dicussion we landed on something slightly better we think.
This yet again more closely matches what many have said about using Lists.

Additionally we keep the ability to edit but we use a different operator
to glue togethor edit functions...'$'...which honestly surprised me but
it works!

In Optics land `a => List b` is similar to a Fold (or a Getter but not quite).
And our Setters are litterally just what Data.Lens calls a Setter.

There are obviouslly some known downsides to using List a in places where
Option a or even just "a" would work but using map and mapPartial you
can still compose those cases into querries and Setters don't have this
issue.

As for reading a result that "should" result in an Option a or "a" value,
you can use expectOne, expectMany, and expectFirst to handle these cases.
Since you would generally have to convert to a Result in such a case any
way. If all else fails you can use require-else to handle anything.

jObjectF and jObjectMap were removed. It makes the jq based examples longer
but argueably more understandable (jObjectMap's type was basically vommit).
I don't think these will be missed.

I also dropeed "convert" because I wasn't really sure anyone was going to
use it and if needed we can bring it back.

Lastly we need to consider the fate of some functions like `at`, `head`,
and `find` to see if/how they're being used. I added findFirst and lookup
which are essentially special (but common) cases of `find`.

If everyone is happy enough with the high level details we can move on to:
- details such as implement, code quality, etc...
- adding features or removing features
- coming up with example problems to test the new library against

After that (or as we go) I can add documentation comments.

I'm going to start thinking about testing now.