[job-cache] Add tool for creating a job cache
This is the start of a tool that will be able to create, retrive, and manage
a cache of jobs. It's missing *many* features that will need to be added to
reach parity with the existing database. This PR only allows you to *add*
things to the database (but the files are copied in). Additionally some important
features like atomicity and allowing multiple job caches to interact at once
are missing but that can be fixed by adding WAL which we will need anyway. Still
this represents a good base that appears like its going to work going forward.