[wcl] Introduce the Wake Core Library
This change just adds an optional type that functions the same way as std::optional more or less.
It doesn't respect assignment quite the same way right now which means that wcl::optional<std::vector<T>>
isn't as efficen't as it could be since if you assign one vector to another that already has capacity,
the current implementation deallocates the first one and reallocates this one despite that not being
needed.