test: Provide dir to bin/ instead of wake as $1
Instead of setting $1 to <something>/bin/wake set it to <something>/bin.
This allows tests to execute any of the tools installed by wake.

Also updates all wake tests to point to bin/wake.

It is mainly needed to support wake-format tests but may be used for
other tools in the future.