vscode: make available in web
Node and web implementations of both client and server are now separated.

Instead of executing the server binary, typescript calls exposed wasm functions.

LSPServer::callMethod() returns the response to a request, as well as diagnostics and a notification (instead of dumping messages to stdout).

The extension is compiled with webpack.