vscode: add stdlib.json (for web)
When running in web, the stdlib packaged with the extension is hosted online and not recognized as a folder by vscode fs api. Thus, it cannot be traversed to find its contents.
So, we generate a list of all stdlib files paths (stdlib.json) and ship it with the extension.