html: Use proper HTML attributes for JSON data element.
`type` is supposed to hold a MIME type, so "wake" is not the right value
to use there. The official MIME type for `<script>` elements that hold
JSON data is supposed to be "application/json".

It looks like we were also using our custom "wake" type in the
JavaScript code to identify the element holding the wake JSON data. This
code was iterating over the entire tree of elements to find the element
with the "wake" type. This is unnecessary, and because there should
always be exactly one element holding that data, we can just give it an
id and search for it by id.