html: Use addEventListner() instead of setting event handler properties.
addEventListner() is more modern, and in cases like these where we're
directly adding event listners in JavaScript, this avoids making the
browser to have to parse the text attribute value since we already have
a handle to the event listener.

This also changes the event handler functions to 1) not be defined on
the `document` object and 2) to not have the `that` parameter. There's
no need to mutate the `document` object to add our event handler
functions, since they can just be free-floating, top-level functions.
Also, the `that` parameter is redundant with the `event` handler in
every single one of these cases, since the `event.target` property will
hold the element that the event fired on.