html: Move all styling to CSS.
Instead of directly setting styles on elements, this moves all the
styles to CSS, and instead the JavaScript now adds or removes CSS class
names to elements whenever we need dynamic styling. This also removes
the dynamically constructed `<style>` element for the use-def
highlighting and replaces it with JavaScript code that queries for all
the uses or the definition and dynamically adds CSS classes.