html: Refactor and modernize HTML/CSS/JS (#1179)
* html: Use proper HTML attributes for JSON data element.

`type` is supposed to hold a MIME type, so "wake" is not the right value
to use there. The official MIME type for `<script>` elements that hold
JSON data is supposed to be "application/json".

It looks like we were also using our custom "wake" type in the
JavaScript code to identify the element holding the wake JSON data. This
code was iterating over the entire tree of elements to find the element
with the "wake" type. This is unnecessary, and because there should
always be exactly one element holding that data, we can just give it an
id and search for it by id.

* html: Add boilerplate HTML tags.

The `<!DOCTYPE html>` declaration is important for opting into modern
day HTML5.

The `<html>`, `<head>`, and `<body>` tags are not strictly required, but
they are there for good form.

The `<title>` tag allows us to declare the title of the page as showed
in the browser tab.

* html: Use addEventListner() instead of setting event handler properties.

addEventListner() is more modern, and in cases like these where we're
directly adding event listners in JavaScript, this avoids making the
browser to have to parse the text attribute value since we already have
a handle to the event listener.

This also changes the event handler functions to 1) not be defined on
the `document` object and 2) to not have the `that` parameter. There's
no need to mutate the `document` object to add our event handler
functions, since they can just be free-floating, top-level functions.
Also, the `that` parameter is redundant with the `event` handler in
every single one of these cases, since the `event.target` property will
hold the element that the event fired on.

* html: Add comments for `update()` function.

* html: Check for null when pressing +/- without hovering over an expression.

In this scenario, `inner` is null, but we will throw an exception trying
to access members of `inner`. Adding the option chaining operator (`?.`)
will allow us to safely fall back to null if the LHS is null.

* html: Use Event.key instead of deprecated .which and .keyCode.

Event.key is the correct modern API to use to get the value of key while
taking into account keyboard locale and layout. This also has the side
effect of making the code more self-documenting.

* html: Move all styling to CSS.

Instead of directly setting styles on elements, this moves all the
styles to CSS, and instead the JavaScript now adds or removes CSS class
names to elements whenever we need dynamic styling. This also removes
the dynamically constructed `<style>` element for the use-def
highlighting and replaces it with JavaScript code that queries for all
the uses or the definition and dynamically adds CSS classes.

* html: Extract colors into CSS custom properties.

This makes it easier to set the palette in one place.