Clean up extra file handles in shim-wake (#1293)
This change cleans up file handles that wake leaks to shim-wake. Ideally we should be using close-on-exec for all file handles we don't intentionally want to be kept...but that doesn't seem to be happening for whatever reason and this is a lot easier than debugging that sort of issue.