Add bulk logging env var for wake (#1392)
This change adds the WAKE_BULK_LOGGING_DIR environment variable. If turned on, wake will emit a copy all logs to that directory, with each file having its hostname, pid, the current day as part of the file name. This ensures that even over NFS, logging should stay atomic. NFS has no append atomicity but because wake is single threaded and the pid and hostname are included things should be atomic.

The day in the file name is mostly so that tools and scripts for parsing logs can pre-filter the logs they want to view based on the time frames without parsing all of the logs. In theory if a multi-day process is running on the same host with the same pid, or a pid gets recycled some days later this would also be midly helpful but those are quite rare cases.

Log cleanup can be done by deleting files with an expiered modtime but that system is not implemented in wake itself for bulk logging like it is for cache logging.