TTL eviction policy (#1455)
This change adds the TTL eviction policy change. It turns out that the LRU eviction strategy has some major issues with it right now. The new default is a 7-day TTL policy. The function of the TTL eviction policy is far simpler and should not suffer from the same complexity issues.