[result] Fix documentation for getWhenFail and getWhenPass methods.
The examples in the documentation showed passing the same arguments and
getting different results. The examples are incorrect because they
forgot to change the arguments to demonstrate the other case. This fixes
the examples and also replaces the type of the other case to more
clearly indicate that it is totally unused.